refactor: moved `dataFetcher` into `/LimitsTable`
- also reorganized imports in affected files
