+++ src/routes/safe/components/Settings/SpendingLimit/RemoveLimitModal.tsx
@@ -15,9 +15,9 @@ import { getWeb3 } from 'src/logic/wallets/getWeb3'
 import useToken from 'src/routes/safe/components/Settings/SpendingLimit/hooks/useToken'
 import { SPENDING_LIMIT_MODULE_ADDRESS } from 'src/utils/constants'
 
-import { SpendingLimitTable } from './dataFetcher'
 import { RESET_TIME_OPTIONS } from './FormFields/ResetTime'
 import { AddressInfo, ResetTimeInfo, TokenInfo } from './InfoDisplay'
+import { SpendingLimitTable } from './LimitsTable/dataFetcher'
 import Modal from './Modal'
 import { useStyles } from './style'
 import { fromTokenUnit } from './utils'
