+++ src/routes/safe/components/Settings/SpendingLimit/index.tsx
@@ -9,8 +9,8 @@ import Row from 'src/components/layout/Row'
 import { safeSpendingLimitsSelector } from 'src/logic/safe/store/selectors'
 import { grantedSelector } from 'src/routes/safe/container/selector'
 
-import { getSpendingLimitData } from './dataFetcher'
 import LimitsTable from './LimitsTable'
+import { getSpendingLimitData } from './LimitsTable/dataFetcher'
 import NewLimitModal from './NewLimitModal'
 import NewLimitSteps from './NewLimitSteps'
 import { useStyles } from './style'
